refactor: -validate and normalize Instagram URLs; fix type value -Fix circular import: move URL parsing into a shared util -Harden URL check to avoid false positives and handle @username -Clear stale importError on success -Avoid overwriting photo with an empty string
